+++ mod-source-record-storage-server/src/main/java/org/folio/services/RecordService.java
@@ -104,4 +104,13 @@ public interface RecordService {
    */
   Future<Boolean> updateSuppressFromDiscoveryForRecord(SuppressFromDiscoveryDto suppressFromDiscoveryDto, String tenantId);
 
+  /**
+   * Deletes records associated with specified snapshot and snapshot itself
+   *
+   * @param snapshotId snapshot id
+   * @param tenantId   tenant id
+   * @return - future with true if succeeded
+   */
+  Future<Boolean> deleteRecordsBySnapshotId(String snapshotId, String tenantId);
+
 }